EVAKA-4161 Moved absence helper to lib-common
