+++ frontend/package.json
@@ -193,7 +193,6 @@
       "./src/citizen-frontend",
       "./src/e2e-playwright",
       "./src/employee-frontend",
-      "./src/employee-mobile-frontend",
       "./src/lib-common"
     ],
     "reporters": [
