--- frontend/src/employee-mobile-frontend/jest.config.ts
@@ -1,9 +1,0 @@
-import type { Config } from '@jest/types'
-
-const config: Config.InitialOptions = {
-  displayName: 'employee-mobile-frontend',
-  preset: 'ts-jest',
-  testEnvironment: 'node',
-  testRunner: 'jest-circus/runner'
-}
-export default config
