+++ frontend/src/employee-mobile-frontend/tsconfig.json
@@ -5,7 +5,6 @@
     "composite": true,
     "jsx": "react",
     "outDir": "../../dist/employee-mobile-frontend",
-    "types": ["jest"],
     "suppressImplicitAnyIndexErrors": true
   },
   "references": [
