+++ frontend/src/lib-common/utils/absences.ts
